HAL: massacre
