+++ applications/gpio/scenes/gpio_scene_usb_uart_config.c
@@ -1,6 +1,6 @@
 #include "../usb_uart_bridge.h"
 #include "../gpio_app_i.h"
-#include "furi_hal.h"
+#include <furi_hal.h>
 
 typedef enum {
     UsbUartLineIndexVcp,
