+++ applications/gpio/views/gpio_usb_uart.c
@@ -1,6 +1,6 @@
 #include "../usb_uart_bridge.h"
 #include "../gpio_app_i.h"
-#include "furi_hal.h"
+#include <furi_hal.h>
 #include <gui/elements.h>
 
 struct GpioUsbUart {
