+++ applications/subghz/subghz_i.c
@@ -485,7 +485,7 @@ uint32_t subghz_random_serial(void) {
     static bool rand_generator_inited = false;
 
     if(!rand_generator_inited) {
-        srand(DWT->CYCCNT);
+        srand(osKernelGetTickCount());
         rand_generator_inited = true;
     }
     return (uint32_t)rand();
