+++ firmware/targets/f7/furi_hal/furi_hal_spi_i.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <stm32wbxx_ll_spi.h>
+
+struct FuriHalSpiBus {
+    SPI_TypeDef* spi;
+    FuriHalSpiBusEventCallback callback;
+    FuriHalSpiBusHandle* current_handle;
+};
