+++ firmware/targets/f7/furi_hal/furi_hal_usb.c
@@ -3,6 +3,7 @@
 #include "furi_hal_usb.h"
 #include <furi_hal_power.h>
 #include <stm32wbxx_ll_pwr.h>
+#include <stm32wbxx_ll_gpio.h>
 #include <furi.h>
 
 #include "furi_hal_usb_i.h"
